cmd/hiveview: use ES modules in webapp
This restructures the JS part of the app to use ES modules. All external
resources are now included, so the app can run without an internet
connection.
