+++ cmd/hiveview/assets/extlib/bootstrap.module.js
@@ -1,0 +1,5 @@
+// This is a module wrapper for loading bootstrap.
+// bootstrap needs jQuery, so we load that first.
+
+import './jquery.module.js'
+import './bootstrap-4.4.1.min.js'
