+++ cmd/hiveview/assets/extlib/jquery.dataTables.module.js
@@ -1,0 +1,4 @@
+// This is a module wrapper for DataTables.
+
+import './jquery.module.js'
+import './jquery.dataTables-1.10.25.min.js'
