+++ cmd/hiveview/assets/images/details_close.png
