+++ cmd/hiveview/assets/images/details_open.png
